build : folder Structure
