+++ .prettierrc.js
@@ -2,6 +2,6 @@ module.exports = {
   bracketSpacing: false,
   jsxBracketSameLine: true,
   useTabs: true,
-  singleQuote: true,
+  dobuleQuote: true,
   trailingComma: 'all',
 };
