+++ __tests__/App-test.js
@@ -5,7 +5,7 @@
 import "react-native";
 import React from "react";
 import renderer from "react-test-renderer";
-import App from "../App";
+import App from "../app";
 
 // Note: test renderer must be required after react-native.
