+++ app/assets/assetsTest.js
@@ -1,0 +1,1 @@
+console.log("asset Test");
