+++ app/utils/utilsTest.js
@@ -1,0 +1,1 @@
+console.log("util Test");
